% Generated by roxygen2 (4.0.1): do not edit by hand
\name{context_plot}
\alias{context_plot}
\title{Creates dotplot with focus highlighted}
\usage{
context_plot(context_data, focus = "none", xlab, title = "",
  state_label = "Washington", colors = portal_colors[c(8, 4, 2)],
  title_size = 1.2)
}
\arguments{
\item{context_data}{data.frame where first columns is a factor for
y-axis and second is numeric for x-axis}

\item{focus}{character corresponding to factor level(s) to highlight}

\item{xlab}{character for x-axis label}

\item{title}{character for plot title}

\item{state_label}{character to match to highlight state}

\item{colors}{colors for the manual color scale}

\item{title_size}{relative size of the title}
}
\description{
Input context_data needs to have column 1 groups and column 2 data
}

