% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_clean.R
\name{cr_clean}
\alias{cr_clean}
\title{Clean up colnames from Portal, allow option to select columns and to 
convert to \code{Date} class.}
\usage{
cr_clean(df, select = NULL, date = TRUE, date.type = 1,
  qry.type = "all.unique")
}
\arguments{
\item{df}{data.frame frame stored procedure}

\item{select}{optional character vector of columns to keep}

\item{date}{boolean indicating whether any \code{datetime} columns}

\item{date.type}{1 is quarter and 2 is year}

\item{qry.type}{should be converted to \code{Date} classes (default is 
\code{T}).}
}
\description{
Clean up colnames from Portal, allow option to select columns and to 
convert to \code{Date} class.
}
