% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_json.R
\name{df_to_json}
\alias{df_to_json}
\title{Export \code{data.frame} to \code{JSON}}
\usage{
df_to_json(df, file, varname = "foods")
}
\arguments{
\item{df}{data.frame to convert}

\item{file}{name of file to save to}

\item{varname}{name of variable to assign to the JSON object}
}
\description{
This is based on the type of JSON file that the original nutrient Parallel Coordinates
was built off of. It seems to be sort of "row oriented", whereas \code{rjson::toJSON}
is "column oriented".
}

