% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/new_utils.R
\name{fac.to.num}
\alias{fac.to.num}
\title{Factor to numeric via character}
\usage{
fac.to.num(x)
}
\arguments{
\item{x}{factor to be converted to numeric}
}
\description{
Factor to numeric via character
}

