% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/factor_to_numeric.R
\name{factor_to_numeric}
\alias{factor_to_numeric}
\title{Factor to numeric via character}
\usage{
factor_to_numeric(x)
}
\arguments{
\item{x}{factor to be converted to numeric}
}
\description{
Factor to numeric via character
}

