% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finish_dashboard_data.R
\name{finish_dashboard_data}
\alias{finish_dashboard_data}
\title{Helper function to finalize the data for the County Dashboard application.}
\usage{
finish_dashboard_data(spark_base, fact_base)
}
\arguments{
\item{spark_base}{The list object with county, region, and state data
needed by the sparklines.}

\item{fact_base}{The list object with county, region, and state data needed
by the fast facts.}
}
\value{
The function returns a list with the data divided into list elements.
* \code{x$year_limits} - vector with the start year and end year extracted
                         from spark_base
* \code{x$county} - list with a dataframe for "context" and one for "data"
* \code{x$region} - list with a dataframe for "context" and one for "data"
* \code{x$state}  - list with a dataframe for "context" and one for "data"
}
\description{
This function takes the results from \code{get_sparklines_data()} and
\code{get_fast_facts_data()} and converts the data to the structure and 
format the County Dashboard application expects.

The final data structure is divided as follows:
- metadata specifying the "window" of time covered by the data (year start
  to year end) - impacts the context for the collection of sparklines in that
  it sets the time axis limits
- county data
 - "context": the values used to create the visualization context (e.g., 
   names, min/max values)
 - "data": the values used to create the visualization contents (e.g., the
   trend lines)
- region data
 - "context"
 - "data"
- state data
 - "context"
 - "data"
}

