% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_county_dashboard_data.R
\name{get_county_dashboard_data}
\alias{get_county_dashboard_data}
\title{Update the data for the County Dashboard application.}
\usage{
get_county_dashboard_data(annie_connection, poc_connection)
}
\arguments{
\item{annie_connection}{Active RODBC connection the annie MySQL server. No
testing done to insure valid connection - should be handled external to this
function.}

\item{poc_connection}{Active RODBC connection the POC SQL server. No
testing done to insure valid connection - should be handled external to this
function.}
}
\value{
Function attempts to verify if the update succeeds but failure is not
terminal to avoid interrupting the update collection.

Returns a simple list. If all update steps seem succesful, first element
is TRUE and second element is character string "Success."

If any update steps seem to fail, first element is FALSE and second element
is a character vector of detected issues.
}
\description{
This is a helper function that pulls together the data needed by the
County Dashboard application (https://github.com/pocdata/county_dashboard),
formats the data as needed, and then writes the data to a folder sharing the
name of the project repo ("county_dashboard").

Two collections of data are produced by \code{get_sparklines_data()} and
\code{get_fast_facts_data()}:
\itemize{
 \item{data to support the trends over time (aka - the sparklines)}
 \item{data to support summary descriptive statistics (aka - the fast facts)}
}

These collections are then split into county, region, and state variants,
along with some crucial context and titling information for each variant,
by \code{finish_dashboard_data()}.

The final data results are:
\itemize{
 \item{county_context.csv (titling and sparklines/fast facts context)}
 \item{county_data.csv (sparklines/fast facts data)}
 \item{region_context.csv}
 \item{region_data.csv}
 \item{state_context.csv}
 \item{state_data.csv}
 \item{year_limits.csv (the min and max year covered by the data)}
}

Data sources used in the above process:
\itemize{
\item{annie: variety of stored procedure calls}
\item{POC: dbo.ref_lookup_county_region}
\item{POC: dbo.ref_lookup_census_population}
\item{POC: public_data.unemployment}
\item{POC: public_data.hs_graduation}
\item{POC: public_data.free_lunch_county}
}
}

