% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_fast_facts_data.R
\name{get_fast_facts}
\alias{get_fast_facts}
\title{Helper function to get and clean County Dashboard "fast facts" data.}
\usage{
get_fast_facts(poc_connection)
}
\arguments{
\item{poc_connection}{Active RODBC connection to the POC SQL server.}
}
\value{
Returns a list with three data frames - county, region, and state - needed
to support the different fast fact views. These data frames need further
processing prior to app consumption.
}
\description{
This function gathers the base data needed for the County Dashboard fast
facts and handles initial cleaning and preparation of this data.

Once finished, the fast facts data - along with the sparklines data
gathered by \code{get_sparklines_data()} - is handed off to
\code{finalize_dashboard_data()} to put it in the final format needed
for app consumption.
}

