% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_site_dashboard_data.R
\name{get_site_dashboard_data}
\alias{get_site_dashboard_data}
\title{Update the data for website dashboards.}
\usage{
get_site_dashboard_data(annie_connection)
}
\arguments{
\item{annie_connection}{A character string for ODBC connection defaulted to \code{"annie"}}
}
\value{
The function returns a json object.
}
\description{
This is a function to update the data for the website dashboards that
should be run after each data load.
}

