% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_dashboard_data_wrapper.R
\name{get_site_dashboard_data_wrapper}
\alias{get_site_dashboard_data_wrapper}
\title{Wrapper for .}
\usage{
get_site_dashboard_data_wrapper(annie_connection)
}
\arguments{
\item{annie_connection}{A character string for ODBC connection defaulted to \code{"annie"}}
}
\value{
The function returns a json object which is placed in a new folder when 
\code{get_portal_app_data()} is run.
}
\description{
This is a function to update the data for the website dashboards that 
should be run after each data load.
}

