% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_sparklines_data.R
\name{get_sparklines_data}
\alias{get_sparklines_data}
\title{Helper function to get and clean County Dashboard trends data.}
\usage{
get_sparklines_data(annie_connection)
}
\arguments{
\item{annie_connection}{Active RODBC connection to one of the annie MySQL
servers.}
}
\value{
Return a single data frame with the base sparklines data. However, further
work needs to be done with this data before it is ready for County
Dashboard consumption. This is handled by \code{update_county_dashbaord()}.
}
\description{
This function gets the data needed for the County Dashboard trends (aka -
the sparklines) and handles cleaning and preparation of the data.
}

