% Generated by roxygen2 (4.0.1): do not edit by hand
\name{\%btwn\%}
\alias{\%btwn\%}
\title{Tests (element-wise) if the first argument falls within
the range of the second.}
\usage{
a \%btwn\% b
}
\description{
Tests (element-wise) if the first argument falls within
the range of the second.
}
\examples{
1 \%btwn\% c(0, 5)
0:4 \%btwn c(0.5, 3, 3.5)
}

