% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/custom_operators.R
\name{\%coalesce\%}
\alias{\%coalesce\%}
\title{Fills in NAs in the first argument from the second}
\usage{
a \%coalesce\% b
}
\description{
Fills in NAs in the first argument from the second
}

