% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h.R
\name{h}
\alias{h}
\title{Shorthand for \code{head()}}
\usage{
h(x, ...)
}
\arguments{
\item{x}{the object to apply head to}

\item{...}{additional arguments to be passed the head helper functions}
}
\description{
Shorthand for \code{head()}
}

