% Generated by roxygen2 (4.0.1): do not edit by hand
\name{layOut}
\alias{layOut}
\title{Arranging ggplots}
\usage{
layOut(...)
}
\arguments{
\item{...}{Each argument should be of the form \code{list(plot, rows, columns)},
where \code{plot} is a ggplot (or similar), and \code{rows} and \code{columns}
are consecutive sequences indicating the row and column numbers for \code{plot}
to span.}
}
\description{
Provides a \code{layout}-like interface for arranging ggplots of different sizes.
}
\examples{
\dontrun{
gg <- ggplot(mtcars, aes(x = hp, y = mpg)) + geom_point()
layOut(list(gg, 1:2, 1:3),
       list(gg, 3, 1:2),
       list(gg, 3, 3))
}
}
\author{
Alan D. Jassby and James E. Cloern (originally from the \code{wq} package).
}

