% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecations.R
\name{Mode}
\alias{Mode}
\title{Likely to be removed in future versions. Avoid dependencies.}
\usage{
Mode(x)
}
\arguments{
\item{x}{Vector for which you want to know the mode.}
}
\description{
Mode of a vector. Found on Stackoverflow.
}
