% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{poc_colors}
\alias{poc_colors}
\title{poc_colors}
\format{An object of class \code{character} of length 4.}
\usage{
poc_colors
}
\description{
Vector of 4 hex colors from POC theme.
}
\keyword{datasets}

