% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deprecations.R
\name{portal_clean}
\alias{portal_clean}
\title{Clean up colnames from Portal, allow option to
select columns and to convert to \code{Date} class.}
\usage{
portal_clean(df, select = NULL, date = T)
}
\arguments{
\item{df}{data.frame frame stored procedure}

\item{select}{optional character vector of columns to keep}

\item{date}{boolean indicating whether any \code{datetime} columns
should be coverted to \code{Date} classes (default is \code{T}).}
}
\description{
Clean up colnames from Portal, allow option to
select columns and to convert to \code{Date} class.
}

