% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_test.R
\name{print.sp_results}
\alias{print.sp_results}
\title{Print method for sp_test.}
\usage{
\method{print}{sp_results}(x, ...)
}
\arguments{
\item{x}{Object to print (inherited from print).}

\item{...}{Additional arguments to pass to print.}
}
\description{
Simple method to insure that the result turned by sp_test only prints out
the concise summary by default. Not exported.
}

