% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{quickfacts}
\alias{quickfacts}
\title{quickfacts (from US Census Bureau)}
\format{An object of class \code{data.frame} with 40 rows and 20 columns.}
\usage{
quickfacts
}
\description{
Data frame characteristics of each county and of the state. 
\code{qf_dict} has the column keys.
}
\keyword{datasets}

