% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sql.R
\name{read_sql}
\alias{read_sql}
\title{Function for reading in a SQL query.}
\usage{
read_sql(filename, silent = TRUE)
}
\arguments{
\item{filename}{You know what goes here}

\item{silent}{boolean on whether you want warnings ignored (TRUE).}
}
\description{
This probably still needs some work.
}
