% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{ref_lookup_gender}
\alias{ref_lookup_gender}
\title{ref_lookup_gender(from test_annie)}
\format{An object of class \code{character} of length 2.}
\usage{
ref_lookup_gender
}
\description{
Data frame with columns for pk_gnder, cd_gndr and tx_gndr.
}
\keyword{datasets}
