% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/remove_leading_blank_lines.R
\name{remove_leading_blank_lines}
\alias{remove_leading_blank_lines}
\title{Removes blank lines from the top of a file}
\usage{
remove_leading_blank_lines(input, output = input)
}
\arguments{
\item{input}{character path to the file}

\item{output}{character path to output file (defaults to overwrite input). Set to \code{NA}
to disable output.}
}
\value{
invisibly returns the modified text
}
\description{
Reads in text (usually from a file), removes any blank lines (or lines containing
only whitespace) from the top of the file, (optionally) writes the modified
file, invisibly returns the modified text.
}

