% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deprecations.R
\name{resetwd}
\alias{resetwd}
\title{Likely to be removed in future versions. Avoid dependencies.
Sets working directory to \code{"S:/Data Portal"}}
\usage{
resetwd()
}
\description{
Likely to be removed in future versions. Avoid dependencies.
Sets working directory to \code{"S:/Data Portal"}
}

