% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rmd_to_md.R
\name{rmd_to_md}
\alias{rmd_to_md}
\title{Creates md config files from Rmd source}
\usage{
rmd_to_md(base_path = getwd())
}
\arguments{
\item{base_path}{character path to directory.
Defaults to the working directory.
Must have a subdirectory called \code{RMarkdown}).}
}
\value{
No value returned.
}
\description{
Works well if your working directory is the root
of a portal content repo. (You can alse give a filepath).
Looks in a subdirectory called \code{Rmarkdown/} for
\code{.Rmd}
files, knits them to \code{.md} and removes any
blank lines from the top.
}

