% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/safe_folder.R
\name{safe_folder}
\alias{safe_folder}
\title{Safely create a uniquely named folder.}
\usage{
safe_folder(target_name = "temp", target_path = ".")
}
\arguments{
\item{target_name}{The folder name the function will try to create.}

\item{target_path}{Where the folder should be created. Defaults to the
current working directory (".").}
}
\value{
The function will return a list with three items.

\code{x$result} will be a boolean. \code{TRUE} on success and \code{FALSE} on
failure.

\code{x$details} will be a character string explaining what occurred.
Probably only interesting on failure.

\code{x$full_path} will be a character string giving the path of the created
folder. Useful for automation where you need to know what folder was
created.

On success, the given folder will also have been created in the target
directory.
}
\description{
When writing scripts that need to write to the local machine, we often
need to create a unique (often temporary) directory.

This function attempts to create a directory of the given name - or "temp"
if no name is given - but will avoid overwriting or recreating an already
existing directory.

If a directory of the given name exists, it will instead attempt to create
a variant of the given name.
}

