% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sp_cr.R
\name{sp_cr}
\alias{sp_cr}
\title{Creates call for stored procedures}
\usage{
sp_cr(sp = c("ia_trends_counts", "ia_trends_rates", "ia_safety",
  "ooh_pit_counts", "ooh_pit_rates", "ooh_flow_entries_counts",
  "ooh_flow_entries_rates", "ooh_flow_exits", "ooh_reentry", "ooh_outcomes",
  "ooh_outcomes_12m", "ooh_outcomes_24m", "ooh_outcomes_3m",
  "ooh_wb_familysettings", "ooh_wb_siblings", "ooh_wb_siblings_pvt",
  "ooh_wb_siblings_uncensored", "population_household", "population_person"),
  db = c("mysql", "sqlserver", "poc", "test_annie"), date = 0, age = 0,
  ethnicity = 0, county = 0, IHS = 0, reporter = 0, access = 0,
  allegation = 0, finding = 0, race = 0, gender = 0,
  intial_placement = 0, longest_placement = 0, length_of_stay = 0,
  number_of_placements = 0, service = 0, budget = 0, dependency = 0,
  year = 0)
}
\arguments{
\item{sp}{stored procedure, default is for ia_trends_counts, use sp_names to find the names of all stored procedures}

\item{db}{choose database to contect to mysql, splserver, poc or test_annie}

\item{date}{}

\item{age}{}

\item{ethnicity}{}

\item{county}{}

\item{IHS}{}

\item{reporter}{}

\item{access}{}

\item{allegation}{}

\item{finding}{}

\item{race}{}

\item{gender}{}

\item{longest_placement}{}

\item{length_of_stay}{}

\item{number_of_placements}{}

\item{service}{}

\item{budget}{}

\item{dependency}{}

\item{year}{}

\item{initial_placement}{}
}
\value{
A call for stored procedures
}
\description{
Creates call for stored procedures
}

