% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecations.R
\docType{data}
\name{sp_names}
\alias{sp_names}
\title{Stored Procedure Names (may be removed - avoid dependencies)}
\format{An object of class \code{character} of length 19.}
\usage{
sp_names
}
\description{
sp_names returns the names of all stored procedures.
}
\keyword{datasets}

