% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sp_test.R
\name{sp_test}
\alias{sp_test}
\title{Tests \code{stored_procedure} functionality on the local machine.}
\usage{
sp_test(connection)
}
\arguments{
\item{connection}{An RODBC connection or a character vector that can be
passed to \code{odbcConnect} to create an RODBC connection appropriate for
your local machine.}
}
\description{
\code{stored_procedure} is the workhorse of \code{pocr}. However, the strings
it produces are only useful if the local machine is correctly configured to
connect to the POC SQL server, the database(s) involved are up to date, and
the stored procedures call by the strings are valid.

\code{sp_test} creates the SQL call strings for all available stored
procedures and tests the results of attempting to use each one to retrieve
data from the POC SQL server.

This function allows you to quickly check if the strings produced by
\code{stored_procedure} can be used to get valid results on the local
machine.

The function will attempt to capture any errors and direct the user to their
likely cause(s).

This error detection is limited. The function will test if:
\itemize{
  \item The provided POC SQL server connection works for accessing the
  needed database(s) and table(s).
  \item Each of the available stored procedures return valid results.
}
}

