% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{theme_clean}
\alias{theme_clean}
\title{A blank ggplot theme, borrowed from somebody on github}
\usage{
theme_clean(base_size = 12, ...)
}
\arguments{
\item{base_size}{defaults to 12, should be fine mostly}
}
\description{
A blank ggplot theme, borrowed from somebody on github
}

