% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_poc}
\alias{theme_poc}
\title{POC theme for ggplot}
\usage{
theme_poc(font = "Open Sans", slant = FALSE, gridlines = c("xy", "x", "y",
  "", "none", NA), expand.margin = FALSE, ...)
}
\arguments{
\item{font}{Partial-matches to "Arial", "PT Sans" (for web) or "Frutiger LT Std 45 Light" (for print).}

\item{slant}{boolean for whether x-axis labels should be slanted. (FALSE)}

\item{gridlines}{What gridlines to include. Deafult is \code{"xy"}. Use \code{"x"} or \code{"y"}
for x or y only, any of \code{""}, \code{NA}, or \code{"none"} will result in no lines.}

\item{expand.margin}{boolean for whether the right margin needs padding. Useful if \code{slant}
is \code{TRUE} and the labels on the right side are lengthy.}

\item{...}{additional arguments passed to theme_bw}
}
\description{
Based off of \code{theme_bw}, this frames the plot in POC blue,
eliminated minor grid lines, and will try to match to a POC font.
There are also some other options.
}

