% Generated by roxygen2 (4.0.1): do not edit by hand
\name{trend_plot_state}
\alias{trend_plot_state}
\title{Plots re-revamped trend charts}
\usage{
trend_plot_state(trend_data, type = c("ooh", "ia", "ihs"), title = NA,
  title_size = 1.2)
}
\arguments{
\item{trend_data}{data frame with **2 columns, in order,**
that are date, count.}

\item{type}{Character, either "ooh", "ia", or "ihs". Used for y-axis label.}

\item{title}{Character. Title to go at top. Defaults to an appropriate "Trends in..."}

\item{title_size}{numeric, relative title size}
}
\description{
Nice barchart of counts, facets on geography.
}

