% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/update_county_data.R
\name{update_county_data}
\alias{update_county_data}
\title{Function for updating county data from SQL Server}
\usage{
update_county_data()
}
\description{
Function for updating county data from SQL Server
}

