% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/update_omf.R
\name{update_ofm}
\alias{update_ofm}
\title{Update the OFM tables in the POC SQL Server database.}
\usage{
update_ofm(poc_connection, start_year = 2000, end_year = 2014)
}
\arguments{
\item{poc_connection}{Need to provide an active RODBC connection to the
POC SQL server.}

\item{start_year}{Optional parameter where the user can specify the start
year.}

\item{end_year}{Optional parameter where the user can specify the end
year.}
}
\value{
The function updates the POC SQL database and returns TRUE on success.
}
\description{
Imports, cleans, and organizes the data needed for the OFM tables in the POC
SQL server database. It then drops and recreates the tables with the cleaned
data.

HOW IT WORKS

We first download the data from the source. We then use a helper function
to import the data as a list of dataframes, one for each xlsx file.
We then filter out irrelevant rows and columns. Since some columns of
the xlsx file represent values, we melt. Moreover, the values of the column
take on values for two separate variables. Thus we mutate to split the
columns. Once the tidying of the data is done, we create another data frame
which initially takes on race-based totals. These totals are used to compute
the calculated other count which is given by:

others = total - hispanic - black - white

We then combine all of the tables and then create rows for 2015. 2015 data
will be replicated from the 2014 data.

Once we have 2014 data
}

