% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/update_ospi_data.R
\name{update_ospi_data}
\alias{update_ospi_data}
\title{Update OSPI data.}
\usage{
update_ospi_data()
}
\arguments{
\item{connection}{A valid RODBC connection to the SQL Server Database. It is
assumed that the connection has been checked with
\code{validate_RODBC_input()} or has been otherwise validated - minimal
testing of the input will be done.}
}
\description{
This function updates the POC SQL Server database with the most current
versions of data extracted from the OSPI website. This includes data
pertaining to high school graduation rates (public_data.hs_graduation)
and free and reduced lunch (public_data.free_lunch_county).

The function simply reads the relevant Excel files from the host website,
formats the data, and then uses the data to drop and re-create the relevant
tables in the POC SQL server database.
}

