% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/validate_config.R
\name{validate_config}
\alias{validate_config}
\title{Validates the json in a config file}
\usage{
validate_config(text)
}
\arguments{
\item{text}{chracter vector (1 line per element) of a Rmd or md config file}
}
\value{
TRUE or FALSE indicating if the json is valid
}
\description{
This takes a character vector of text from a config file
(such as that produced by readLines) and validates the json
}
\details{
We assume that config files
\itemize{
    \item optionally begin with a single R code chunk on the first line
    \item there is no content before the json starts other than the single
    optional code chunk
    \item the first non-json line following the json begins with a \code{#}
}
}

