% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/validate_RODBC_input.R
\name{validate_RODBC_input}
\alias{validate_RODBC_input}
\title{Validating requests for an RODBC connection.}
\usage{
validate_RODBC_input(connection)
}
\arguments{
\item{connection}{An RODBC connection or character vector that can be
passed to odbcConnect to create an RODBC connection.}
}
\value{
The function will return \code{TRUE} if an RODBC object is provided or
can created, along with with an appropriate success message.

If an RODBC object is not provided or cannot be created, \code{FALSE} will
be returned with an appropriate failure message.
}
\description{
This is a helper function used to insure users provide valid input where
they need to specify a valid RODBC connection.

Here 'valid' means either an object of class RODBC or a character string
that - when passed to \code{RODBC::odbcConnect()} - creates such an RODBC
object.
}

