% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write_test.R
\name{write_test}
\alias{write_test}
\title{A simple test to see if R has write permissions.}
\usage{
write_test(test_directory = ".")
}
\arguments{
\item{test_directory}{The directory you want to test against. Defaults to
"." for the current working directory.}
}
\value{
The function returns a boolean value \code{TRUE} on success (indicating
R has write permissions in the test directory). On failure, the function
returns a message articulating where the attempt to write failed.
}
\description{
This simply attempts to create a folder in the named directory (the
working directory by default). If successful, the test folder is deleted.

To minimize the risk of overwriting or deleting an existing folder, the
test folder is created using \code{safe_folder()}. As a result, the test
may fail under the very unlikely circumstance that folders named "temp1"
to "temp15" exist in the test directory. The user will be notified should
this occur.
}

