% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/county_to_office.R
\name{county_to_office}
\alias{county_to_office}
\title{Takes a county (string or numeric) and gives the corresponding rows of
ref_lookup_office}
\usage{
county_to_office(county)
}
\arguments{
\item{county}{string or numeric county_cd}
}
\description{
Takes a county (string or numeric) and gives the corresponding rows of
ref_lookup_office
}

