% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{df_to_json}
\alias{df_to_json}
\title{Export \code{data.frame} to \code{JSON}}
\usage{
df_to_json(df, file, varname = "foods")
}
\arguments{
\item{df}{Data frame to convert.}

\item{file}{name of file to save to.}

\item{varname}{name of variable to assign to the JSON object.}
}
\description{
This is based on the type of JSON file that the original nutrient Parallel Coordinates
was built off of. It seems to be sort of "row oriented", whereas \code{rjson::toJSON}
is "column oriented".
}

