% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecations.R
\name{fac.to.num}
\alias{fac.to.num}
\title{Deprecated. Use \code{factor_to_number}.}
\usage{
fac.to.num(...)
}
\arguments{
\item{...}{All input passed to new function.}
}
\description{
Deprecated. Use \code{factor_to_number}.
}

