% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_dashboard_data.R
\name{get_site_dashboard_data}
\alias{get_site_dashboard_data}
\title{Get dashboard data.}
\usage{
get_site_dashboard_data(con)
}
\arguments{
\item{con}{A ODBC connection}
}
\value{
The function returns a json object.
}
\description{
This is a function to update the data for the website dashboards that
should be run after each data load.
}
