% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_sparklines_data.R
\name{get_sparklines_data}
\alias{get_sparklines_data}
\title{Helper function to get and clean County Dashboard "sparklines" data.}
\usage{
get_sparklines_data(annie_connection, year_start = NA, year_end = NA)
}
\arguments{
\item{annie_connection}{Active RODBC connection to one of the annie MySQL
servers.}

\item{year_start}{Optional. Can specify the start year as a character
vector. If NA, the data sources will be assessed for a smallest minimum.}

\item{year_end}{Optional. Can specify the end year as a character vector.
If NA, the data sources will be assessed for a largest maximum.}
}
\value{
Returns a list with three data frames - county, region, and state - needed
to support the different trend line views. These data frames need further
processing prior to app consumption.
}
\description{
This function gathers the base data needed for the County Dashboard trend
plots (aka - the sparklines) and handles initial cleaning and preparation
of this data.

Once finished, the sparklines data - along with the fast facts data
gathered by \code{get_fast_facts_data()} - is handed off to
\code{finalize_dashboard_data()} to put it in the final format needed
for app consumption.
}

