% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_operators.R
\name{\%coalesce\%}
\alias{\%coalesce\%}
\title{coalesce operator}
\usage{
a \%coalesce\% b
}
\arguments{
\item{a}{object with NAs to fill}

\item{b}{object with NA replacement material}
}
\description{
Fills in NAs in the first argument from the second
}

