% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pocr.R
\docType{package}
\name{pocr}
\alias{pocr}
\alias{pocr-package}
\title{POC Utility Functions}
\description{
This package collects a variety of functions and variables constructed
for use by POC staff work with POC data. It includes official POC and
portal colors, POC themes, functions for efficiently querying POC databases,
and more.
}
\details{
See especially \code{\link{stored_procedure}}.
}

