% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{portal_colors}
\alias{portal_colors}
\title{portal_colors}
\format{An object of class \code{character} of length 12.}
\usage{
portal_colors
}
\description{
Vector of 12 hex colors used in the portal
}
\keyword{datasets}
