% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{qf_dict}
\alias{qf_dict}
\title{Key to \code{quickfacts}}
\format{An object of class \code{data.frame} with 18 rows and 3 columns.}
\usage{
qf_dict
}
\description{
Verbose descriptions of \code{quickfacts} columns.
}
\keyword{datasets}

