% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{quickfacts}
\alias{quickfacts}
\title{quickfacts (from US Census Bureau)}
\description{
Data frame characteristics of each county and of the state.
\code{qf_dict} has the column keys.
}

