% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{quickfacts}
\alias{quickfacts}
\title{quickfacts (from US Census Bureau)}
\format{\preformatted{'data.frame':	40 obs. of  20 variables:
 $ fips     : int  53000 53001 53003 53005 53007 53009 53011 53013 53015 53017 ...
 $ text     : chr  "Washington" "Adams" "Asotin" "Benton" ...
 $ PST045212: int  6897012 19005 21888 182398 73687 71863 438287 3995 101996 39350 ...
 $ PST120212: num  2.6 1.5 1.2 4.1 1.7 0.6 3 -2 -0.4 2.4 ...
 $ AGE135211: num  6.5 11.2 5.8 7.4 7 4.6 6.7 4.8 6.2 7.2 ...
 $ AGE295211: num  23.2 35.2 21.3 26.7 24.9 17.8 26 19 23.9 26.7 ...
 $ RHI125211: num  82 92.5 94.8 91.5 94.6 88.3 88.1 95.1 92.2 94.1 ...
 $ RHI225211: num  3.8 1.1 0.5 1.5 0.5 1 2.1 0.4 0.8 0.7 ...
 $ RHI325211: num  1.8 4 1.5 1.2 1.6 5.3 1 1.5 1.8 1.8 ...
 $ RHI425211: num  7.5 1.1 0.7 2.9 1 1.5 4.3 0.6 1.6 1 ...
 $ RHI525211: num  0.7 0 0.2 0.2 0.2 0.2 0.7 0.3 0.3 0.2 ...
 $ RHI625211: num  4.3 1.2 2.3 2.7 2.1 3.6 3.8 2 3.3 2.2 ...
 $ RHI725211: num  11.6 59.9 3.1 18.8 26.5 5.3 7.8 6.2 8.1 28.9 ...
 $ RHI825211: num  72.1 38.1 92.2 74.2 69.9 84.3 81.4 89.7 85.3 67.4 ...
 $ POP815211: num  17.8 50 3.8 17 22.5 7.4 13.5 3 7.5 26.7 ...
 $ EDU635211: num  89.8 67.5 88.7 88.2 82.7 91.4 90.8 87.5 86.5 80.1 ...
 $ HSD310211: num  2.5 3.17 2.38 2.69 2.62 2.23 2.66 2.19 2.52 2.69 ...
 $ PVY020211: num  12.5 24.5 14.6 12.2 12.6 13.6 11.7 15.4 17.5 16.7 ...
 $ LND110210: num  66456 1925 636 1700 2921 ...
 $ POP060210: num  101.2 9.7 34 103 24.8 ...
}}
\usage{
quickfacts
}
\description{
Data frame characteristics of each county and of the state.
\code{qf_dict} has the column keys.
}
\keyword{datasets}

