% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{"ref_lookup_raceeth_census"}
\alias{"ref_lookup_raceeth_census"}
\alias{ref_lookup_raceeth_census}
\title{"ref_lookup_raceeth_census" (from test_annie)}
\format{An object of class \code{data.frame} with 13 rows and 2 columns.}
\usage{
ref_lookup_raceeth_census
}
\description{
Data frame with columns for cd_race_census and tx_race_census.
}
\keyword{datasets}
