% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deprecations.R
\name{read.sql}
\alias{read.sql}
\title{Deprecated. Use \code{read_sql}.}
\usage{
read.sql(...)
}
\arguments{
\item{...}{All input passed to new function.}
}
\description{
Deprecated. Use \code{read_sql}.
}

