% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/new_utils.R
\name{read.sql}
\alias{read.sql}
\title{Function for reading in a SQL query.}
\usage{
read.sql(filename, silent = TRUE)
}
\arguments{
\item{filename}{You know what goes here}

\item{silent}{boolean on whether you want warnings ignored (TRUE).}
}
\description{
This probably still needs some work.
}

