% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{ref_lookup_county}
\alias{ref_lookup_county}
\title{ref_lookup_county (from test_annie)}
\description{
Data frame with columns for the county_cd, the text county
and the regional code cd_region. Taken from SQL Server table of the same
name. I've ordered the entries so that row number corresponds to county_cd
for the actual counties 1:39.
}

