% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{ref_lookup_county_region}
\alias{ref_lookup_county_region}
\title{ref_lookup_county_region (from test_annie)}
\format{An object of class \code{data.frame} with 40 rows and 7 columns.}
\usage{
ref_lookup_county_region
}
\description{
Data frame with columns for the county_cd, the text county
and the regional code cd_region. Taken from SQL Server table of the same 
name. I've ordered the entries so that row number corresponds to county_cd 
for the actual counties 1:39. Additionally, has columns for region, the
most useful of which are probably region_6_cd, region_6_tx, and small_fl.
}
\keyword{datasets}

