% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{ref_lookup_office}
\alias{ref_lookup_office}
\title{ref_lookup_office (from test_annie)}
\description{
Data frame with columns cd_office, tx_office,
cd_office_county_grp, and tx_office_county_grp. Taken from SQL Server table
of the same name.
}

