% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{ref_lookup_office}
\alias{ref_lookup_office}
\title{ref_lookup_office (from test_annie)}
\format{An object of class \code{data.frame} with 46 rows and 4 columns.}
\usage{
ref_lookup_office
}
\description{
Data frame with columns cd_office, tx_office, 
cd_office_county_grp, and tx_office_county_grp. Taken from SQL Server table 
of the same name.
}
\keyword{datasets}
