% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deprecations.R
\name{sp_cr}
\alias{sp_cr}
\title{Deprecated. Use \code{stored_procedure}.}
\usage{
sp_cr(...)
}
\arguments{
\item{...}{All input passed to new function.}
}
\description{
Deprecated. Use \code{stored_procedure}.
}

