% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deprecations.R
\docType{data}
\name{sp_names}
\alias{sp_names}
\title{Stored Procedure Names (may be removed - avoid dependencies)}
\format{\preformatted{ chr [1:19] "ia_trends_counts" "ia_trends_rates" "ia_safety" ...
}}
\usage{
sp_names
}
\description{
sp_names returns the names of all stored procedures.
}
\keyword{datasets}

