% Generated by roxygen2 (4.0.1): do not edit by hand
\name{theme_poc}
\alias{theme_poc}
\title{POC theme for ggplot}
\usage{
theme_poc(font = c("Arial", "PT Sans", "Frutiger LT Std 45 Light"),
  slant = FALSE, gridlines = TRUE, expand.margin = FALSE, ...)
}
\arguments{
\item{font}{Partial-matches to "Arial", "PT Sans" (for web) or "Frutiger LT Std 45 Light" (for print).}

\item{slant}{boolean for whether x-axis labels should be slanted. (FALSE)}

\item{gridlines}{boolean for whether there should be any gridlines (TRUE)}

\item{expand.margin}{boolean for whether the right margin needs padding. Useful if \code{slant}
is \code{TRUE} and the labels on the right side are lengthy.}
}
\description{
Based off of \code{theme_bw}, this frames the plot in POC blue,
eliminated minor grid lines, and will try to match to a POC font.
There are also some other options.
}

