% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_plot.R
\name{trend_plot}
\alias{trend_plot}
\title{Plots trend charts}
\usage{
trend_plot(trend_data, type = c("ooh", "ia", "ihs"), title = NA,
  title_size = 1.2, font = "Open Sans")
}
\arguments{
\item{trend_data}{data frame with **3 columns, in order,**
that are date, count, geography.}

\item{type}{Character, either "ooh", "ia", or "ihs". Used for y-axis label.}

\item{title}{Character. Title to go at top. Defaults to an appropriate 
"Trends in..."}

\item{title_size}{numeric, relative title size}

\item{font}{Character, named font to use.}
}
\description{
Nice barchart of counts, facets on geography.
}
