% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_RODBC_input.R
\name{validate_RODBC_input}
\alias{validate_RODBC_input}
\title{Validating requests for an RODBC connection.}
\usage{
validate_RODBC_input(connection)
}
\arguments{
\item{connection}{An RODBC connection or character vector that can be
passed to odbcConnect to create an RODBC connection.}
}
\value{
The function returns a list with three elements. 

\code{x$test_result} will return boolean \code{TRUE} if an RODBC object is 
provided or was successfully created - it will be \code{FALSE} otherwise.

\code{x$test_message} returns a character string giving a more verbose
description of the success or fail. It can be inspected for details or
used to convey success/failure information to the user.

\code{x$connection} returns the created/validated connection. This is how 
the connection can be retrieved for further use.
}
\description{
This is a helper function used to insure users provide valid input where
they need to specify a valid RODBC connection.

Here 'valid' means either an object of class RODBC or a character string 
that - when passed to \code{RODBC::odbcConnect()} - creates such an RODBC
object.
}
